Fixing Strings unit tests, moving Typolib Strings methods in their own class
